[REF] account_reports: ignore session boolean
Account reports save options into the session so users don't lose them
when reloading the page.To prevent some reports to use those session
options we used the 'ignore_session' variable.

The 'ignore_session' could have 3 values:
- 'read' to prevent the reading of session options
- 'write' to prevent the writing of session options
- 'both' to prevent reading and writing of session options

The thing is that there is not real usage for those 3 cases. In
practice, the session options are completely ignored (read and write) or
never ignored. The distinction between 'read' and 'write' is then
useless.

This all means that we can simplify that variable by making it a boolean
 instead.

closes odoo/enterprise#46128

Signed-off-by: Olivier Colson (oco) <oco@odoo.com>