[FIX] social: show missing helper on applying Group By
Before this commit:
Open Social with no account connected. Select the 'By Stream' filter from
Group By. We see an empty screen.
We have to do this change following the changes introduced in the file
`kanban_controller.js` in the PR below:
https://github.com/odoo/odoo/pull/114024

Technical reason:
https://github.com/odoo/enterprise/blob/master/social/static/src/js/stream_post_kanban_renderer.js#L50-L52
The method `get showNoContentHelper()` in the code above, returns
`!model.hasData()` which has the model value as `KanbanSampleModel`.
The `hasData()` returns `true` when the group by is applied even with empty data.
And hence the issue.

We now use the method `useModelWithSampleData` to pass the model value
as `StreamPostKanbanModel`.

After this commit:
We can see the 'No Stream Added yet!' helper.

task-3468029

closes odoo/enterprise#47275

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>