[FIX] planning: correct pill alloc.hours when grouped by
Issue:
When grouping by 2 factors, in some circumstances, the total
allocated hours for the group is not equal the sum of the allocated
hours of all the pills in said group. This happens when we have an
open shift which happens to have some part of it's shift outside the
company working hours, which shouldn't be the case, as open shifts
are 24/24.

Steps to reproduce:
- Install Planning
- Planning > Group by Role then Resource
- Choose an open shift, and set the end time to 20h
- Notice the aggregate of the allocation hours for the group is not
  correct.

Cause:
There is an inconsistency of the allocated hours of the record, and
what is computed as alloc.hours on the pill after it had been
enriched. The main condition which split the computation between if
we take the calendar into question, or not, `isFlexibleHours`,
returns `undefined` when the `resource_id` is `false`, which is the
case for open shifts, as they have no resource assigned to them.
This leads to the condition to be evaluated falsy, taking the code
path where the calendar is taken into account.

Fix:
This patch introduces multiple small corrections:
- Explicit casting to boolean for the return value of
  `isFlexibleHours`, to correct the signature return value.
- Add an additional condition specifically for open shifts, they
  should take the same code path as shifts which resource is a
  flexible resource.
- Do not multiply twice the allocated hours with the alloc.%, it's
  already taken into account in the allocated hours of the record.

Affected versions:
saas-16.2 up to master

Reference:
opw-3438537

closes odoo/enterprise#46251

X-original-commit: 807b2a4b2ec6e752a3ed5b3edc03292ca5345360
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>