[FIX] mrp_workorder: start WO when a QC is done
Once a WO's Quality Check is done, if the WO is still not started, it
will be started (without launch the timer).
It makes more sense for a WorkOrder to be started as soon one of its
stap is processed.

Also, before this commit, we had a bug where QC Component Registration's
quantity was 0 instead of the quantity to do. It was because the QC's
defaut quantity is computed when the previous QC is done, and since this
computation depends of the WO's producing quantity and its WO was not
started (so, producing qty at 0), the QC's qty was computed to 0 too.

task-3456817

closes odoo/enterprise#46701

Related: odoo/odoo#134075
Signed-off-by: Arnold Moyaux (arm) <arm@odoo.com>