[FIX] mail_enterprise: rendering of emails in chatter
Before this commit, messages of type email were not rendered
properly in the chatter. For example, the email may display
text with dark color even when in dark theme or the layout of
bootstrap elements like `btn` might not look correct.

These style issues are caused by combining the style in content
of the email (e.g. inline styles) with the style of the webclient,
giving the impression that the original visual of the email is buggy.

This commit fixes the issue by showing a slightly transformed
style of email messages in chatter that is readable and looks nice
with the Odoo theme at hand. In particular with dark theme, the
background color matches the bubble color and the text is white.

Sometimes the exact visual of the original email is desired.
A button "Show Original Email" in the top-right corner of the
message bubble allows seeing the visual of email in its original
intention.

Task-3437069

closes odoo/enterprise#46849

X-original-commit: 2cf58a9a29c382449b1bc79686b7365698d3ea5d
Related: odoo/odoo#134234
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>