[FIX] sign: stopiteration error when attachments_iter is empty
This error occurs when the next() function has no more items to retrieve from
the iterator and we receive an empty `attachments_iter` when calling the next()
function.

To produce this error turn on the developer mode and directly open the
`sign.template` model `form view` from the `Open View` and click on the save
button without entering anything.

Error: StopIteration: null

To handle this issue we have added a try-except block. using that if we do not
get an `attachment_id` or `name`, the record will not be saved.

sentry-4410298400

closes odoo/enterprise#46650

X-original-commit: bfa2a5be477867933148a0c4091cb8afb9553195
Signed-off-by: Achraf Ben Azzouz (abz) <abz@odoo.com>