[IMP] l10n_mx_edi: Improve cancellation flow
Add a cancellation wizard allowing the user to choose a reason for cancellation.

In Mexico, the cancellation reason '01' is particular. For this flow, you need to create and sign
first the invoice that will replace the original invoice. For that flow, the users need to use the
"Create Replacement Invoice" button first, then choose the cancellation reason '01' on the request
cancel wizard.
For sale orders, we need to ignore the replacement invoice until the original invoice is well cancelled.

Task: 3102918
Part-of: odoo/enterprise#46805