[ADD] l10n_at_saft: SAF-T file export for Austria
The Austrian Bundesministerium für Finanzen tax office may request a
SAF-T (standard audit file for tax) report on a on-demand basis.
This is generally ahead of a VAT audit.

Currently the SAF-T export for Austria is not implemented.
This commit adds a new module to export the SAF-T for Austria.

The following information is currently exported:
1. The data that is available through the account_saft module.
2. The mapping of the exported accounts to the chart of accounts given in the
   Austrian SAF-T specification.

task-3354661

Part-of: odoo/enterprise#47405