[IMP] pos_*: skip receipt screen blinking
Currently when the pos.config is set to "Skip Preview", when printing a
receipt, the receipt or bill screen shows very quickly and disappears.
This is not good UX.

Now, the receipt is printed from the payment screen itself. We don't
need to show the receipt screen at all.

closes odoo/enterprise#46187

Taskid: 3441238
Related: odoo/odoo#132550
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>