[IMP] sale_subscription: Sale Subscription Plans
This commit aims to improve the usability of subscription by
changing the behaviour of template and recurrence.

Before this commit subscription "so.template" were used after
they had been assigned which was antithetical to what a template
should be. On top of that, subscription and rental share
"sale.temporal.recurrence" but had completly different use and
paradigm for it and it's related pricings.

This commit aims to reduce the complexity and chaos related to those issue
by removing fields that were use after the creation of the subscription
and the billing recurrence realted fields on a new,
subscription-related model "sale.subscription.plan".

This plan will now regroup all the information related to the
subscription using it and any change to this model will be
applied to all related subscriptions. This make subscription
more user friendly for user managing multiple similar subscriptions.

In addition, invoicing journal was removed from subscription and
merged into sale directly. (See community PR : #TODO)

Task-id: 3471665
Part-of: odoo/enterprise#46621