[FIX] website_helpdesk: rename website helpdesk menu when un/pub teams
Issue:
By default, when a website has one published helpdesk team, the
label of the helpdesk menu is "Help". When we add a second team on
the website, each team gets it's own menu, and the label should
reflect the name of the team. Currently the first team, which had
the label "Help" is not renamed to the team name when we publish
another team on the website. The inverse process should also be
true: if we have multiple teams, and we remove all of them but one,
said team menu label should be renamed to Help, which is currently
not the case.

Steps to reproduce:
- Install Helpdesk and Website
- Publish one of the helpdesk team to the website with "Website Form".
- Notice it has the has the label "Help" on the website menu.
- Publish another team, refresh the website
- Notice the first label is still "Help", while the second is the
  name of the newly published helpdesk team.

Cause:
In `_ensure_website_menu(self)`, we set the name of the of the newly
written to team, creating the correct label for the new menu that we
are adding, but there is no routine to update the names of the other
team's menu linked to the website we are modifying.

Fix:
With the teams provided from the read_group, we scan for those that
whose menu may need a label update based on the count of teams we
have on the website.

Affected versions:
saas-15.2 up to master

Reference:
opw-3453794

closes odoo/enterprise#46597

X-original-commit: 1a6f5d7fcf69bfeddb65f0abe733eaa1a46aa6f9
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>