[FIX] website_helpdesk: add missing _lt
See community counterpart.
Note that in this commit, the `_t()` was added by commit [1] but never
worked as it was supposed to be `_lt()`. But adding `_t()` at least made
the terms detected as translated terms, meaning that despite never being
really translated in the UI, those were exported to transifex and there
is translation inside `.po` files. Just replacing the `_t()` by `_lt()`
will load the translation.

[1]: https://github.com/odoo/enterprise/commit/dd095897a27b45abce52ceb39b287212de198311

opw-3471873

closes odoo/enterprise#46590

X-original-commit: 95ce3839ec363aaffb75af939c4f5351b4677f3a
Related: odoo/odoo#133692
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>
Signed-off-by: Romain Derie (rde) <rde@odoo.com>