[IMP] mrp_workorder: shop floor use WO quantity for register production
In register production steps on WO cards in the shop floor, the quantity
of the MO was used as the total quantity to produce. However, there are
cases where the WO quantity deviates from the MO quantity.
This commit now uses the remaining_qty from the WO instead.

Task: 3456827 (nr 4)
Part-of: odoo/enterprise#48379