[FIX] sale_subscription: remove base.group_portal access to sale.temporal.recurrence
Before this commit, the sale.temporal.recurrence records had security rules for
portal users. It was only needed to display the list of sale.order in the
/my/subscription page. We decide to remove this portal access and to use
sudo when necessary. The /my/subscription/<int:order_id> is using the sudoed
record, therefore no change is needed in that view.

closes odoo/enterprise#46234

Taskid: 3221012
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>