[FIX] spreadsheet_edition: forbid insert in spreadsheet groupby date in rows and cols
Issue:
=====
When you select groupby date month in rows and groupby date year in cols
in pivot view it still shows  the button insert in spreadsheet.

Steps to reproduce the error:
=============================
- Go to sales/reporting -> pivot view
- Groupby rows by month
- Groupby cols by year
- insert in spreadsheet is enabled

Origin of the issue:
====================
When we check on duplicates we don't consider the the fact that
date:month and date:year are the same if applied in rows and cols at the
same time. it's okay to have them both on only rows or cols. Having this
behavior will result in error when displaying the spreadsheet because in
`pivot_model.js` we check only on the name of the groupby which is date
to classify a groupby is a col or a row . (https://github.com/odoo/odoo/blob/cc3e899be84f44a8d6b1161ef9ae278e8f75c1c3/addons/spreadsheet/static/src/pivot/pivot_model.js#L561)

Solution:
=========
As dicussed with LUL, I added a check on the date to make sure it exist
in only rows or cols but not both.

opw-3390240

closes odoo/enterprise#46242

Signed-off-by: Lucas Lefèvre (lul) <lul@odoo.com>