[ADD] room: manage rooms bookings
PURPOSE:
--------

The purpose of this module is designed to streamline the management of meeting
rooms within organizations. The significance of this module arises from the need
to efficiently allocate and utilize meeting spaces, which are vital for various
business activities.

SPECIFICATION:
--------------

Backend:

- Room Configuration:

In the backend, the "Rooms" module offers comprehensive room management
capabilities. When creating a room, room managers can input essential details
such as the room name, available amenities, and associate it with a specific
company and office. Each room is provided with a dedicated URL for frontend
access, and room managers can customize this URL that includes a unique access
code. Additionally, room images for frontend representation and the ability to
set colors for the room's website are integral features.

Rooms view is facilitated through both kanban and list views, offering efficient
management options. The kanban view provides quick insights into room
availability, displaying whether a room is currently in use or available, along
with the start time of the next meeting if available.

A convenient search panel allows users to filter rooms based on company name and
office name.

- Meeting Scheduling:

The backend functionalities of the module extend to meeting management, ensuring
seamless scheduling, updates, and removal of meetings. Various views, including
Gantt view, kanban view, calendar view, and list view, are available for
efficient meeting management. When scheduling meetings from the backend, users
are required to provide essential information such as the meeting title,
start time, and end time.

Moreover, the module incorporates a chatter feature within the form view of
meetings. This chatter function tracks critical meeting details, including the
meeting's start and end times, and any changes made to the organizer's name.
This feature enhances transparency and provides an audit trail for meeting-
related activities.

Frontend:

- Current rooms status

The frontend view is intended to be used on a static device (such as a tablet)
next to the corresponding room. It shows the current time and the availability
of the room. If the room is busy, it additionally displays the remaining time
for the current meeting. it also shows all the upcoming meetings in the
sidebar.

- Quick Schedule Meetings:

Users can quickly schedule meetings from the frontend with minimal input. They
can book the room for the next 15, 30, and 60 minutes if the room is available.

- Future Schedule:

Users can also schedule meetings that will start later. The scheduling process
begins by selecting the desired date for the meeting. Users can then select the
start and end time slots for the meeting.

- Edit Meetings:

Users have the option to reschedule or delete meetings as needed.

Users can edit meetings by clicking on them in the sidebar. In that case the
user can change the start and end dates as well as the name. They can also
delete a meeting by clicking on the fa-close icon of the booking card in the
sidebar.

TECHNICAL SPECIFICATION

- Models

room_booking model:

This model is dedicated to managing bookings. It includes fields for titles,
start times, end times, organizers, and potentially other relevant
meeting-related information.

room_room model:

The room_room model is designed for the management of rooms. It contains fields
such as name, amenities, office, room background image, and colors for the
frontend view.

In order for the frontend view to be updated in real time, the model sends
messages through the bus service (create/delete/update bookings, update the
room).

room_office model:

This model represents offices within a company.

- Access Rights

Members of the "Room Manager" group possess comprehensive access rights. They are
granted the ability to: create, update, and delete rooms.

Internal users can create, update, and delete every booking.

Public users can access the frontend view using the url of the room (that contains
a unique short code). They can read, create, update and delete the bookings of a
room through a controller, provided that they have the access token for the room.
This access token is used when making requests from the frontend view.

Task-3414166

closes odoo/enterprise#48095

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>