[MOV] (website_)knowledge: move frontend files to website_knowledge
Purpose:
--------
Since the refactoring of knowledge for portal users [1], the "frontend"
view is only relevant when website is installed (portal users don't use
it anymore, they use the same one as internal users).

This commit moves the assets and methods related to this frontend view
from knowledge to website_knowledge.

Task-3346593

[1]: https://github.com/odoo/enterprise/commit/d66829ca7fcbf25bd4683963908d0973734abf96

Part-of: odoo/enterprise#43304