[FIX] mrp_mps: skip moves out of dates range
If the actual date of an outgoing stock move is out of the mps dates
range, a traceback will appear

To reproduce the issue:
(Need `sale_management`)
1. In Settings:
   - Enable Multi-Step Routes
   - Configure MPS:
     - Range: Daily
     - Number: 7
2. Create a product P
3. Add P in MPS
4. Inventory > Configuration > Rules, edit "Pull from Stock to Custo":
   - Lead Time: 10 days
5. Create and confirm a SO with P:
   - Delivery Date: today + 15 days
6. Open MPS

Error: an Odoo Server Error appears: "IndexError: list index out of
range"

When confirming the SO, a procurement has been created for "today +
15 days". Then, the delivery rule has been used, its lead time has
been applied: the SM is created with a date equal to "today + 5
days", which is inside MPS dates range. As a result, when getting
the outgoing SM for the MPS:
https://github.com/odoo/enterprise/blob/c679d196ac165a8e62aefd2fd5fc69e103c6ebdd/mrp_mps/models/mrp_mps.py#L891-L903
In `_get_moves_and_date` we find the SM and cancel the rule lead
time (so we get the actual expected date): we have a tuple with the
SM and "today + 15 days". But here is the issue: in the while-loop,
we try to find which range index corresponds to that date -> it is
not possible as we are out of dates ranges -> index will be greater
than `len(date_range)` -> `IndexError`

Note: Similar issue could happen with an incoming SM

OPW-3372579

closes odoo/enterprise#46627

X-original-commit: 39cb1941ec75a808475c2f2674ed520e1616c9e2
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Adrien Widart (awt) <awt@odoo.com>