[FIX] sale_subscription: add domain for non-recurring data
Before this commit, all the sales quotation status was set to cancel, but the
data still appeared in reports and non-recurring filters in both graph and pivot
views.

in this commit, add domain for cancelled records to not showing in reporting
graph view and pivot view.

task-3500508

closes odoo/enterprise#48473

X-original-commit: 76a16df7a7be2b140894fe913c7e526fdb897a49
Related: odoo/odoo#137735
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>