[FIX] sale_subscription: add domain for non-recurring data
Before this commit, all the sales quotation status was set to cancel, but the
data still appeared in reports and non-recurring filters in both graph and pivot
views.

in this commit, add domain for cancelled records to not showing in reporting
graph view and pivot view.

task-3500508

closes odoo/enterprise#46687

Related: odoo/odoo#133920
Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>