[PERF] account_accountant: hit trigram index in `_retrieve_partner`
Issue:
Clicking on "Reconcile X items" from the main dashboard in our internal
database can take upward to a minute to open the view.

Analysis:
When we are preparing the auto balance lines, we are retrieving the
partner_id, which is a compute calling `_retrieve_partner`. If we
didn't find a partner from the statement line or the bank account,
we are searching for the partner based on his name. This domain uses an
`ilike`, which would benefit from the presence of a trigram index on
said field. Sadly this is not the case on our production database,
but we have already an index on `display_name`, as it's already a
search criteria in the `_rec_names_search` of the `res.partner`
model. Adding another trigram index on `name` would be too heavy as
the table is already really large, and `display_name` contains
`name`, and in this use-case, they are most likely equal, as the
partner we are looking for don't have a parent.

Solution:
Change the domain criteria to be based on the `display_name` instead
of the `name`.

Results:
On our test database, the search
```py
self.env['res.partner'].search([
        (X, '=', 'some_partner_name'),
        ('company_id', '=', False),
        ('parent_id', '=', False)],
    limit=1)
```
takes:
- *Before* (X = `'name'`): we were hitting an unrelated index (a btree
  index on display_name + id) and then filtering based on the name.
  This was churning through 48GB of data, and taking in the ballpark
  of **19s** to complete
- *After* (X = `'display_name'`): we are hitting the trigram index
  first, then sorting on the result row instead, only taking **360ms**.

A **50x** speed improvement.

Reference:
task-3479202

Notes:
In saas-16.4, 'display_name' on res.partner had been renamed/replaced
by 'complete_name'. See commit: odoo/odoo@3c62ca1eb96d571b2b686b5caee370324c589ab4

closes odoo/enterprise#46507

X-original-commit: 5a49a850727075a19985e715fe44f3e8f72d9711
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Piryns Victor (pivi) <pivi@odoo.com>