[FIX] l10n_ec_edi: get discount total of calculated discounts per line
To calculate the sum of the discounts and obtain the total for the EDI, the discount was recalculated considering the unit price. This calculation is wrong when we have a tax with the configuration of 'included in the price'.

We fixed the calculation of the total discount using the '_l10n_ec_prepare_edi_vals_to_export_USD' method that returns the discount per line, in this way in the EDI both the total discount value matches the values in the lines

closes odoo/enterprise#46361

Signed-off-by: Josse Colpaert <jco@odoo.com>