[IMP] web_enterprise: badges color contrast
This commit add the `tags_list.dark` and `colorlist.dark` files to 
override the tags colors for the dark_mode. We chose to "swap" the color
and bg-color logic with slight adjustments to have a lighter text on a 
darker background for the dark mode tags.

The colorlist colors had to be adapted to display the proper background
colors of the new tags. 

task-3201038

Part-of: odoo/enterprise#49666
Co-authored-by: chgo-odoo <chgo@odoo.com>