[ADD] pos*: new module to send order number to preparation display
pos*: pos_order_tracking_display, pos_preparation_display,
pos_self_order_preparation_display

This commit adds a new module.
This module is a bridge module between the preparation display and the
kiosk mode of the pos_self_order module. It sends the order to the
preparation display when the order is done. It also adds the
tracking number of the order on the up-right corner of the order-card
in the preparation display and an indication if this order is to
take-away or to eat-in.

closes odoo/enterprise#47203

Task-id: 3495477
Related: odoo/odoo#134980
Signed-off-by: David Monnom (moda) <moda@odoo.com>