[IMP] hr_payroll_account: Allow payment registering on payslips
PURPOSE
=======

When you work with Payroll, the flow is to generate payslips through batches
and advance them to Done state. From the Done state, you'll have option to set
them as paid. For this, you need to have the option "generate a sepa file" in the
payroll option.

On a payslip the button "mark as paid" will update the status to "paid". On a batch,
there will be "Mark as paid" like the payslip, this will only change the status.

Or, there is a "Create Payment Report" that'll generate a SEPA.xml file with all
information to pay employees.

If we don't work with SEPA files we need to be able to register a payment made
from Accounting.

SPECIFICATION
=============

When you click on register payment on a payslip that is not included in a batch

The payment wizard should be similar to the one in expense or accounting

closes odoo/enterprise#45994

Taskid: 3318545
Related: odoo/odoo#132495
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>