[FIX] web_studio: render custom report images as raw data
The sale order report (and other reports) generated in the portal page
does not render the images that could have been added with Studio

Steps to reproduce:
1. Install Sales and Studio
2. Go to Sales and open any quotation
3. Toggle Studio and add an image field anywhere in the form
4. Go to reports and edit "Quotation / Order"
5. Add the new image field in the report and close Studio
6. Create a quotation for Joel Willis, add an image and a product
7. Send the quotation by mail
8. Open the sent mail (Settings > Technical > Email > Emails)
9. Open the sale order link in a new incognito tab
10. Download the sale order with the button at the top left of the page
11. The report doesn't contain the image

Solution:
Add the attribute `qweb_img_raw_data` on binary fields created in Studio
in order to render the images as raw data directly in the html

Problem:
When generating the html that will be used to render the pdf, the image
field is rendered with a src url. This url will then be resolved in the
content_image route but the user is public so he has no rights on the
field. Hence, the placeholder image is returned.

opw-3100989

closes odoo/enterprise#46625

X-original-commit: cecbbc558732585388c7b61a9755122380e0525c
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Guillaume Merlin (megu) <megu@odoo.com>