[FIX] knowledge: fix onchange of input selection controlling visibility
Currently, the input selection controlling the visibility of the article
in the workspace is not properly updated whenever the user selects a new
value. The user has to reload the page to see the change.

Steps to reproduce the error:
1. Create an article in "Workspace".
2. Open the share panel of the article you created.
3. Change the visibility of the article.
=> The input selection changes but returns to its initial state.

TO BE: The input selection should be updated and reflect the new
configuration of the article.

task-3482257

closes odoo/enterprise#46411

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>