[FIX] web_studio: use correct boolean atttribute serialization
The view editor in web_studio was refactored in [commit 1]. A new OWL component
(Property) now encapsulates different types of properties on views or their
nodes. Those properties are usually set as an attribute on the corresponding
xml node in the view.

Before this commit, these values were serialized on the client side before
being sent to the server by the rpc service. However, now the raw values are
being sent over by the different sidebar editor classes resulting in python-
like values ending up in the view xml.

The bug described in the associated ticket arises because in some places
archParseBoolean is being used to interpret the value (not triggering the bug),
while in other places JSON.parse is being used (triggering the bug).

Patching the edit_view route to translate boolean attributes into their JSON
counterparts avoids the bug and results in the same xml as before the refactor.

opw-3473254

[commit 1]: https://github.com/odoo/enterprise/commit/d6d8aeff58e46373f72fc932cd84822a29ea76fb

closes odoo/enterprise#47791

X-original-commit: 88fa94c1fa0f0f7ce6531021783441a885b56edc
Signed-off-by: Bastien Fafchamps (bafa) <bafa@odoo.com>
Signed-off-by: Tom De Caluwé (tdc) <tdc@odoo.com>