[FIX] mrp_workorder : access shop floor with company without employees
Issue:
======
When you have multiple companies and switch from a company to another
one where no employee is registered and then access shop floor app
will raise an error.

Steps to reproduce the error:
=============================
- Install Manufacturing
- Switch to My Company (chicago)
- Try to access Shop floor

Origin of the error:
====================
When you connect with the first company the employee_id will be added
to the `EMPLOYEES_CONNECTED` session , and when you switch to the second
company he will stay there and that's why we have the employee from the
first company will be in the session and thus raises an error because
he is not an employee of this company.

Exptected Behavior:
===================
- WOs should still show and not be clickable when the user is just a user
and not employee.
- No WOs should show when there are no employees that can be signed in.

opw-3470707

closes odoo/enterprise#48119

X-original-commit: 5d18b5e6ddf407f5ece63233ab93fd0317915cdb
Signed-off-by: Tiffany Chang (tic) <tic@odoo.com>
Signed-off-by: Mahdi Cheikh Rouhou (macr) <macr@odoo.com>