[IMP] pos*: unify numpads
pos*: pos_iot, pos_restaurant_preparation_display

At the moment, the point_of_sale module has 3 different implementations of
numpads. This is absurd. In this PR we unify them all into a single
numpad component.

One of the reasons why there are multiple numpads is because each requires
a slightly different set of buttons. The new `Numpad` component provides
a simple API which allows the consumer to specify what the buttons should be.

This is done by proving the `buttons` prop, which has the following type:
```js
buttons: {
	type: Array,
	element: {
		type: Object,
		shape: {
			value: String,
			text: { type: String, optional: true },
			class: { type: String, optional: true },
			disabled: { type: Boolean, optional: true },
		},
	},
}
```

In this PR we also provide a simple set of helper functions for the tours
and adapt all of the existing tests accordingly.

closes odoo/enterprise#46427

Task: 3470346
Related: odoo/odoo#132245
Signed-off-by: Joseph Caburnay (jcb) <jcb@odoo.com>