[FIX] l10n_pe_edi: Fix IAP external tests
Since https://github.com/odoo/odoo/pull/122663, calls to IAP services
are prevented by default in tests, and raise an AccessError.

As such, the IAP l10n_pe_edi tests were failing in the external_l10n
nightly build - see for example
https://runbot.odoo.com/runbot/build/49918376

The solution is to remove the testing flag on the current thread when
calling iap_jsonrpc(). (See the solution implemented in the website
module in the above PR.)

closes odoo/enterprise#46591

X-original-commit: 39c4da12041f5dc9698fc14895845fdf75f5f213
Signed-off-by: Josse Colpaert <jco@odoo.com>
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>