[IMP] website_helpdesk: avoid searching empty string in help center
In the help center, it is possible to click on the magnifier button next to the search bar
without typing anything in the search bar, which doesn't make sense and can be a bit confusing.

This commit disables the button while the search bar is empty.

Task-3475459

Part-of: odoo/enterprise#46454