[FIX] website_knowledge: reintroduce "load more" test
Purpose:
--------
In the commit of the sidebar refactoring [1], the "load more" test
(introduced in commit [2]) has been removed. The tour has been adapted,
but the related python test has not been moved to website_knowledge (it
has been deleted).
This was an error. Therefore, this commit reintroduces this test.

Task-3346593

[1]: https://github.com/odoo/enterprise/commit/b55ee8f06d7990fc71ee2430029775d5e8277020
[2]: https://github.com/odoo/enterprise/commit/3c0a4d91c31ca435d96e27f0064c0bfc151c07ad

Part-of: odoo/enterprise#43304