[FIX] spreadsheet_edition: use named placeholders for translation
In the english sentence, the user locale is first and the spreadsheet locale
is second.

But it doesn't mean it's the same in every language. Some language have other
grammatical rules (think of Dutch where the verb is last)

By using named placeholder, translators are able to change the order
of the placeholder.

Task: 3483938
Part-of: odoo/enterprise#46460