[FIX] sale_ebay: unable to set NOT NULL on column 'ebay_currency'
Currently, When the user changes the company currency with deactivating the
default currency and try to install sale_ebay module he receives an exception
at the backend.

To reproduce the issue:
1. Create a database without demo data.
2. Install invoicing module.
3. In Settings > Open Companies from Users&Companies menu.
4. Open My Company and change its currency from default to any other currency.
5. Go to Invoicing Settings > click on the currencies button and deactivate the
default currency.
6. Now install sale_ebay module and the exception will occur at the backend side

To solve this issue the 'ebay_currency' field is set as a required field from
XML side.

sentry-3933754992

closes odoo/enterprise#46890

X-original-commit: 39d352bb34cdcfae4b1e51a6aa934f9ee253cc45
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>
Signed-off-by: Saurabh Choraria (sauc) <sauc@odoo.com>