[FIX] hr_contract_reports: give access to contract start date information
In the Contract reporting view, we have access to integers fields:

- # Departure Employee
- # New Employees

but this information is just giving us for the employee of a given
period that are new or departed, but we have no easy way to know which
employee this concerns.

If someone wanted to see which employee are new for a given month, there
is no easy way from this view => the information is available in the
SQL view, but not made available in the model fields.

This commit proposes to make the field available, so the information is
more easily accessible (eg. you could group by "Date First Contract
Started" to get the contract start of the employees).

opw-3444718

closes odoo/enterprise#49275

X-original-commit: 3e896163ef09fcc6574e9cbe2b0f41f60ee1be1c
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>