[IMP] spreadsheet_edition: improve locale mismatch tooltip
Purpose
-------
Currently, if you have a spreadsheet with en_US locale but the user has the
fr_FR locale, a small globe warning appears on top of the spreadsheet.
When hovering on the globe, this message appears:

```
Difference between user locale (fr_FR) and spreadsheet locale (en_US):
- current Date Format: dd/mm/yyyy
- current Number Format: 1 123 456,89
```

Several issues with this text:

- "current" doesn't mean anything. Does it mean "current user" or
  "current spreadsheet" ?
- the displayed formats are the user formats. But what we actually want to
  show is the spreadsheet format! Such that the user knows how to write dates and numbers in the spreadsheet
- Why are the 2nd and 3rd words upper-cased, but not the first ? I'd only
  uppercase the first one or none

Specification
-------------
change the tooltip text to:

Difference between user locale (fr_FR) and spreadsheet locale (en_US). This spreadsheet is using the formats below.
- dates: mm/dd/yyyy
- numbers: 1,123,456.89

notice the displayed formats are the spreadsheet formats, not the users'

Task: 3483938
Part-of: odoo/enterprise#46460