[FIX] appointment: get available slots with different timezones
Issue:
======
When the user has a meeting as allday and the user is in a timezone
different than `UTC` then getting available slots will produce wrong
results.

Steps to reproduce the error:
=============================
- install calendar + appointment + website
- go to calendar and clear all meetings first
- go to to online appointments and create an online appointment
  with only one user (mitchel admin) and having in schedule only monday
  from 2 to 23, select America/Chicago timezone in options
- update the timezone of the user in user setting prefrences to the same
  one too.
- Go back to calendar , create a meeting with allday activated and
  choose a day in the next week a in monday.
- Go to website now and go to appointment and choose the timezone the
  same as before. you will see that that monday will still have some
  slots available.

Origin of the issue:
====================
When calculating availabilities in the backend we use UTC time for the
slots , so some of them might be in the next day which is tuesday so
there is no allday event in that day and in that way it will be
available.

Solution:
=========
I updated the check to check on allday with the user timezone so we
make sure that it will be unavailable on an allday meeting.

opw-3439545

closes odoo/enterprise#48854

X-original-commit: e0f87617bd155253ab59a5f02eb0d3b550464a1a
Signed-off-by: Jérémy Hennecart (jeh) <jeh@odoo.com>