[REF] mail: move imp of `service.insert()` to `Model.insert()`
In preparation to clean code of models, e.g. by simplifying model
insertion and reducing side-effects in record insertion.

This commit moves the implementation of the `insert()` in discuss
services to the models, and drop support of the service methods
`insert()`. The only way to insert a model is through the store
entry, e.g. `store.Thread.insert()`.

closes odoo/enterprise#46499

Related: odoo/odoo#133509
Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>