[IMP] sale_ebay: currency symbol editing without debug mode
Description of the issue/feature this PR addresses:

Prior to this, when multiple currencies with the same symbol where to appear
in a same document or be sent from a country to another with the same currency
symbol, the only information the recipient had was
the currency symbol leading to lack of clarity.

Desired behavior after PR is merged:

This set of two PR has for objective to give the user the possibility to edit
the symbol curency that is displayed everywhere so that if he feels like the
documents and views are lacking clarity, he can improve it.

This was already an option before but only when the debug mode was active.

This commit has the purpose moving the "Use on eBay" option in the
currency form view so that the displayed symbol can be
added to the right column.

Adding this, the "Use on eBay" option is moved to the left column.

task-3484335

closes odoo/enterprise#46527

Related: odoo/odoo#133554
Signed-off-by: Quentin De Paoli <qdp@odoo.com>