[FIX] sale_subscription: set default stage if none on confirm
There are sale order quotations already created which do not have any stage id assigned to them. So when we add a recurring product to that order, we cannot confirm the order. We click on the confirm button and nothing happens to the status or stage of order. Because the sale order does not have any stage id, it prevents it from progressing to the next stage.

Steps to reproduce:
1. Click on an already created quotation and add a recurring product to the product list.
2. Set the expiry date and add a recurrence.
3. Click on confirm button.

Current Behavior:
Nothing happens on clicking the confirm button.

Expected Behavior:
The quotation should move to sale order stage.

OPW-3439674

closes odoo/enterprise#46534

Signed-off-by: Arnaud Joset (arj) <arj@odoo.com>