[FIX] web_enterprise: fix module icon with reduced motion
=== ISSUE ===

With the new design, we introduced a new way layout for the navbar.

This new layout includes the module icon + the module name, while we used
to simply display a nine dots icon in the past. While using Odoo on small
devices, the module icon would become the old good 9 dots icon.

This new layout introduced a friction for user with a reduced motion
option activated on their computer. It would display the module icon plus
the 9 dots icons. This would result in a weird visual render, even if it
does not really cause any accessibility issue.

=== AFTER ===

We define a behaviour for our animation when `reduced-motion` is
activated. To do so, we define some of the style before the media-query
to not write it twice.

task-3473413
part of task-3326263

closes odoo/enterprise#47771

X-original-commit: b62090daa6797e5709a654cbb93df65cc6935bc3
Signed-off-by: Pierre Paridans (app) <app@odoo.com>