[FIX] project_timesheet_forecast: remove undesired tuples in domain
Before this commit, the read_group done in grid view of timesheets
to find additional data with planning.slot could fail because we
use the whole domain used in the search view. That domain has to
be filtered to just have the field available inside of planning.slot
since the domain in the search is exclusively used for
`account.analytic.line` model.

This commit uses `removeDomainLeaves` to remove all tuples in that
domain containing another fields than `project_id`, `employee_id`,
and `user_id` to be sure the domain used for the read_group done
for `planning.slot` model only uses fields defined inside
`planning.slot`.

closes odoo/enterprise#46586

X-original-commit: 34aa2dd2faee59a8cc46a6e90704248a66167b03
Signed-off-by: Xavier Bol (xbo) <xbo@odoo.com>