[FIX] account_sepa_direct_debit: Update SDD tests for new version.
Problem
---------
A new version of the SEPA Direct Debit Schema (Pain 008.001.08) is
available and will soon be usable in some countries. In Belgium for
example, it will be supported by ISABEL soon, and is already published
in the Netherlands.

Objective
---------
Test the validity of the XML generated for the new SEPA version.

Solution
---------
1. Add some postal information to the companies since this is required by
the updated SEPA version (namely city name and country).
2. Test the validity of the XM generated for some companies.
3. Check that an error is thrown when mandatory postal information is
missing.

**Note** This commit also replace the already present `open` by
`file_open` and `get_resource_path` by `file_path`

task-3440881

closes odoo/enterprise#48081

Signed-off-by: Quentin De Paoli <qdp@odoo.com>