[FIX] l10n_fr_fec_import: trim xmlid for accounts
When importing a FEC file, we set the xmlid of the
accounts with the original code given in the file.
Therefore we can't retrieve the xmlid if an other
line have the same code with extra 0s.

Let's assume a line with account code 5120010, then
the xmlid will be set as `l10n_fr_fec_import.account_5120010`
as the account itself is saved with code 512001.
Now if another line has an account code 512001, then
we don't recreate an xmlid because account has already been
created, but when saving the moves related to 512001, we're
looking for xmlid `l10n_fr_fec_import.account_512001` which
doesn't exist.

With this commit, we adapt the way we set the xmlid by
setting it with the account code truncated at 6 digits,
and we do the same when looking for a xmlid in the partner
and move generator. We've also adjusted the tests to ensure
consistent behavior in line with expectations.

opw-3455682

closes odoo/enterprise#46950

X-original-commit: 7dae238631e53f65b85addafa21e8340db1fb989
Signed-off-by: de Wouters de Bouchout Jean-Benoît (jbw) <jbw@odoo.com>
Signed-off-by: Guillaume Vanleynseele (guva) <guva@odoo.com>